requirements.txt
gitpython>=3.1.0
click>=8.0.0
pyyaml>=6.0
